+++ LeetCode/PalindromeNumber.cpp
@@ -15,7 +15,7 @@ public:
 			arr.push_back(part);
 			x /= 10;
 		}
-		int s = 0, e = arr.size() - 1;
+		int s = 0, e = (int)arr.size() - 1;
 		while (s < e)
 		{
 			if (arr[s] != arr[e])
